+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 1. [#24](https://github.com/influxdata/influxdb-client-python/issues/24): Added possibility to write dictionary-style object
 1. [#27](https://github.com/influxdata/influxdb-client-python/issues/27): Added possibility to write bytes type of data
+1. [#31](https://github.com/influxdata/influxdb-client-python/issues/31): Added support for delete metrics  
 
 ### API
 1. [#28](https://github.com/bonitoo-io/influxdb-client-python/pull/28): Updated swagger to latest version